+++ lib/flutter_quill_delta_from_html.dart
@@ -1,6 +1,7 @@
 library flutter_quill_delta_from_html;
 
 export 'package:flutter_quill_delta_from_html/parser/font_size_parser.dart';
+export 'package:flutter_quill_delta_from_html/parser/indent_parser.dart';
 export 'package:flutter_quill_delta_from_html/parser/line_height_parser.dart';
 export 'package:flutter_quill_delta_from_html/parser/html_to_delta.dart';
 export 'package:flutter_quill_delta_from_html/parser/colors.dart';